\name{calculate.cluster.size.moments.from.model}
\alias{calculate.cluster.size.moments.from.model}
\title{Calculate cluster size moments from model}
\usage{
  calculate.cluster.size.moments.from.model(sampleTime,
    sampleStates, maxTime = NA, minTime = NA,
    timeResolution = 50, discretizeRates = FALSE,
    fgyResolution = 100, integrationMethod = "adams")
}
\arguments{
  \item{sampleTime}{}

  \item{sampleStates}{}

  \item{maxTime}{}

  \item{minTime}{}

  \item{timeResolution}{}

  \item{discretizeRates}{}

  \item{fgyResolution}{}

  \item{integrationMethod}{}
}
\description{
  Calculate cluster size moments from model
}

